feat(api): moved injected dependencies for workflow queue service
