+++ apps/api/src/app/events/usecases/add-job/add-digest-job.usecase.ts
@@ -7,10 +7,10 @@ import { AddJob } from './add-job.usecase';
 import { DigestFilterSteps } from '../digest-filter-steps/digest-filter-steps.usecase';
 
 import {
+  CreateExecutionDetails,
   CreateExecutionDetailsCommand,
-  DetailEnum,
-} from '../../../execution-details/usecases/create-execution-details/create-execution-details.command';
-import { CreateExecutionDetails } from '../../../execution-details/usecases/create-execution-details/create-execution-details.usecase';
+} from '../../../execution-details/usecases/create-execution-details';
+import { DetailEnum } from '../../../execution-details/types';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 import { EventsDistributedLockService } from '../../services/distributed-lock-service';
