+++ apps/api/src/app/events/usecases/message-matcher/message-matcher.usecase.ts
@@ -20,13 +20,13 @@ import { SubscriberEntity, EnvironmentRepository, SubscriberRepository, StepFilt
 import { IFilterVariables } from './types';
 import { FilterProcessingDetails } from './filter-processing-details';
 
-import { CreateExecutionDetails } from '../../../execution-details/usecases/create-execution-details/create-execution-details.usecase';
+import {
+  CreateExecutionDetails,
+  CreateExecutionDetailsCommand,
+} from '../../../execution-details/usecases/create-execution-details';
 import { SendMessageCommand } from '../send-message/send-message.command';
 import { EXCEPTION_MESSAGE_ON_WEBHOOK_FILTER } from '../../../shared/constants';
-import {
-  CreateExecutionDetailsCommand,
-  DetailEnum,
-} from '../../../execution-details/usecases/create-execution-details/create-execution-details.command';
+import { DetailEnum } from '../../../execution-details/types';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 
 const differenceIn = (currentDate: Date, lastDate: Date, timeOperator: TimeOperatorEnum) => {
