+++ apps/api/src/app/events/usecases/store-subscriber-jobs/store-subscriber-jobs.usecase.ts
@@ -7,11 +7,11 @@ import { StoreSubscriberJobsCommand } from './store-subscriber-jobs.command';
 
 import { AddJob } from '../add-job/add-job.usecase';
 
-import { CreateExecutionDetails } from '../../../execution-details/usecases/create-execution-details/create-execution-details.usecase';
 import {
+  CreateExecutionDetails,
   CreateExecutionDetailsCommand,
-  DetailEnum,
-} from '../../../execution-details/usecases/create-execution-details/create-execution-details.command';
+} from '../../../execution-details/usecases/create-execution-details';
+import { DetailEnum } from '../../../execution-details/types';
 import { ApiException } from '../../../shared/exceptions/api.exception';
 
 @Injectable()
