+++ apps/api/src/app/execution-details/execution-details.controller.ts
@@ -1,11 +1,11 @@
 import { ClassSerializerInterceptor, Controller, Get, Query, UseGuards, UseInterceptors } from '@nestjs/common';
 import { ApiTags, ApiOperation, ApiOkResponse } from '@nestjs/swagger';
 import { IJwtPayload } from '@novu/shared';
+
 import { UserSession } from '../shared/framework/user.decorator';
 import { JwtAuthGuard } from '../auth/framework/auth.guard';
 import { ExternalApiAccessible } from '../auth/framework/external-api.decorator';
-import { GetExecutionDetails } from './usecases/get-execution-details/get-execution-details.usecase';
-import { GetExecutionDetailsCommand } from './usecases/get-execution-details/get-execution-details.command';
+import { GetExecutionDetails, GetExecutionDetailsCommand } from './usecases/get-execution-details';
 import { ExecutionDetailsResponseDto } from './dtos/execution-details-response.dto';
 
 @Controller('/execution-details')
