+++ apps/api/src/app/execution-details/usecases/create-execution-details/index.ts
@@ -1,0 +1,2 @@
+export { CreateExecutionDetails } from './create-execution-details.usecase';
+export { CreateExecutionDetailsCommand } from './create-execution-details.command';
