+++ apps/api/src/app/execution-details/usecases/get-execution-details/index.ts
@@ -1,0 +1,2 @@
+export { GetExecutionDetails } from './get-execution-details.usecase';
+export { GetExecutionDetailsCommand } from './get-execution-details.command';
